+++ CHANGELOG.md
@@ -20,6 +20,7 @@ in your sub-class.  This method should be used to map keys to values (if necessa
 which maps keys to `null` values.
 - **Breaking** Change `Enum::fromValue($val)` has been renamed to `Enum::keyForValue()`
 - **Breaking** Change: removed `Enum::flip()`
+- **Breaking** Change `Enum::constantMap()` to `Enum::namesAndKeys()`
 
 # 1.1.0
 - **Breaki